core: erase non-written bits before calling I/O rw functions
None of the instructions do it (e.g. mov.w rm, @rn sends the whole 32
bits of rm) because most of the time the writes land on chunks and the
high bits are not read. However, this leads to I/O functions specifying
e.g. access size 2 and getting values with non-zero high words, which is
weird. Clean that up.

(No bug occurred because of this yet. I just saw it randomly.)