gui: move emulation to another thread
This is considered experimental for now. I debugged a few deadlocks
caused by me blatantly using the pthreads wrong, but nothing else since
then, so I'm hopeful it works correctly.

Performance-wise, the immediate benefits are a bit underwhelming.
Boson X goes from like 47 FPS to 50 FPS on the end screen. But it'll be
easier to improve from here in any case. I have yet to check profiler
traces to see if the emulation thread works as smoothly as I hoped.

The effects are better in the browser, where I go from 4-5 FPS to about
7-8 FPS (with variations in each section).