record: support hardware encoder detection
Use Shotcut's trick that invokes `ffmpeg` and tries to generate a fake
frame using the selected encoder: if the request is successful, then the
encoder is available. This is not the best scenario, but it works for now.

Also note that I have changed each `vec*()` to `glm::vec*()` to avoid error
during the CMake build that indicates that this type is ambiguous since
declared both by Azur and vector headers.