core: auto-generate illegal slot exception checks
The exception checks are now generated by gen-isa.py in the wrappers.
We don't want them in the instructions' semantics because in the state
machine this check will be performed statically when recompiling blocks
and will be removed from semantics.

This is made possible by a new attribute !illslot in the ISA. Because
ldc and ldcl are only !illslot when using SR, ldc_sr and ldcl_sr have
been split off, requiring a new tag !overload to gracefully resolve the
(intended) opcode conflict.

Also simplify the table generator.