core: replace delay slot target mechanism with nextPC
This reduces the state tracked during delay slots. Previously, we'd
remember both that we're in a delay slot (the boolean) *and* the jump's
target address. Now, we only need the boolean. This reduces the amount
of work done by mq_cpu_cycle() and will help transitioning to a state
machine.

The new mechanism uses a new variable nextPC to compute in advance the
location of the next instruction. This is precomputed as PC+2 and all
normal instructions end with PC = nextPC instead of PC += 2. The new
assignment is more versatile: in a delay slot, nextPC is modified by
the branching instruction so that the PC = nextPC of the delay slot
instruction performs the actual jump.

As a tiny optimization, PC is no longer incremented to match the
address of delay slot instructions, since PC-relative instructions are
all illegal in delay slots.

Hopefully this mechanism can also simplify management of the DSP loop.