core: rework handling of DSP loop
New bug added: value of RC is decremented *before* last instruction
instead of after, leading to wrong value if stc sr, rm (or equivalent)
is used as last instruction in DSP loop. We can fix this later by
flagging that we're the last instruction and generate the value of SR
dynamically when reading PC.

Old bug fixed: DSP loop will now trigger if the last instruction of the
loop is the delay slot of a delayed branch instruction.

But mostly this removes a path that would run after every instruction
(using the existing path that runs before), and makes "normal"
instructions and delay slots more similar. The only differences now are
syscall invocations and exceptions.