core: fix nasty bug if interrupt hits after bt/bf
These are the only instructions that de-sync cpu->pc and cpu->nextPC
by not ending with cpu->pc = cpu->nextPC (or ending with a delay slot
that will itself do it).

This would cause a problem if background processes ran immediately
after a bt/bf, as then the incorrect value of nextPC would be used in
the exception handler to set SPC.

One fix could be to set nextPC = PC before running background
processes, but I think it makes more sense to keep them synchronized.