core: split common fields off of SR
This allows removing a lot of get/set functions that would just modify
bitfields. It does require reads of SR to go through a special function,
which isn't much of a problem.

This is mostly cosmetic, except for RC, which will be faster to use in
DSP loops. I'll also simplify the checks further by avoiding RE in
favor of a directed precomputed PC-looping value that becomes -1 when
there's no DSP loop.