gui: move record output path to a utility + fix record being invisible
Bit overkill, but the whole updating logic and cache in the recorder
felt wrong; instead I wanted to minimize the API surface of the ffmpeg-
dependent units.

The point now is that there are no filename data in mqRecord, and I'll
keep chipping away at it by removing screenshots next.