ffmpeg: disable dynamic PTS
Disable dynamic PTS by default to allow the recording of fixed fps
content, which is the default for all codecs and containers (mp4,
webm,...). This change implies that now we encode frame 60 times per
second (the frequency of the GUI refresh), which has a huge(?) impact
on the CPU usage (need more investigation).

I also changed the ffmpeg backend to use `bool` prototype instead of `int`
with complex "errno" handling which have simplified a lot of code that
uses the backend.