watch: rework/fix inotify event handling
In order to prepare the game folder synchronisation, I have discovered
that the inotify event reading works only for files with simple events,
but breaks with directory or more complex events that contain file
name information.

This patch reworks the way that we read events to support any type of
event. This will be useful with directory monitoring.