ci: speedup web build by not cloning `ffmpeg` third-parties
A big part of build jobs is to clone ffmpeg which can take up to 2
minutes during bad network timing. Since the web job does not build or
use `ffmpeg`, I have added an option in the Azur script to manually
skip cloning any ffmpeg-related third-party (`nv-codec-header`, `x264`
and `ffmpeg`).

This reduce by ~26% the job time for the web build.